Add files via upload
I was lucky to be part of the project "Euporia", by CoPhiLab-ILC-CNR Pisa, annotating the corpus of Vth cent BCE Athenian tragedies. 
Euporia is a digital annotation platform based on DLS (Domain Specific Languages), a tool thought for reserarchers who do not  have a solid background on DH, but understand the value and utility of digital instruments in humanities. 

Euporia LanGT (Landscapes in Greek tragedies) aims at annotating the intersections between landscapes and religious notions, showing how tragic poets built up some 'imaginary topographies', made of interactions and relations between space (as a socialised dimensione), dramatic actions/words, and divine powers. The latters can be either on scene (as a deus ex machina), but more frequently alluded to in words by tragic language, so that the Athenian spectators could build up the 'sense' of dramatic spaces in their own theatrical experience. 

Euporia LanGT is based 3 main tags, and is structured into triples, in order to be as close as possible to the language of the (future) ontology I am goning to create. 
My DSL is very simple: 
with #ontology I am marking a statement which is always 'true', and recognizable by the cpu even into the entire corpus. 
with #evidence, I am looking at the specificities of a single passage in a single tragedy. This time, relations will be purely contextual. 
with #asociation, I am finally affording the hermeneutic. Where I wish to point a out a relation between palce-action-divine power, there I use #association. 
There are also two additional tags: @bib and #scene, useful to mark up - respectively- bibliographical references justifying a particular choice, and scenic objects related to spaces I analyse.